Added error checking to organizer and event controllers

+++ backend/controllers/organizerController.js
@@ -1,54 +1,49 @@
 const { database } = require('../firebase-admin/index');
 
 
-const createOrganizer =async (req, res)=> {
-    const { description, events, followers, name } = req.body;
+const createOrganizer = async (req, res)=> {
+    const { description, events, followers, name, email } = req.body;
     const orgData = req.body;
   //check for empty fields
-  if(!description||!events||!followers||!name)
+  if(!description||!events||!followers||!name||!email)
   {
    res.status(500).json({
-success: false,
-error: 'One or more fields are missing'
+        error: 'One or more fields are missing'
     });
   }
     database.collection('Organizers').add(orgData).then(() => {
         res.status(200).json({
-            success: true
+            id: orgData.id
         })
     }).catch((error) => {
         res.status(500).json({
-            success: false,
-            error: 'Failed to create organizer'
+            error: error
         })
     });
 };
 
-const readOrganizer=async (req, res)=>{
+const readOrganizer = async (req, res)=>{
     const {id} = req.params;
     const orgRef = database.collection('Organizers').doc(id);
     const orgDoc = await orgRef.get();
     
     if (!orgDoc.exists) {
-        console.log('user not found');
         res.status(404).json({
-            success: false,
-            error: 'User not found'
+            error: 'Document does not exist'
         })
     } else {
         res.status(200).json(orgDoc.data())
     }
 };
 
-const readOrganizerByEmail=async (req, res)=>{
+const readOrganizerByEmail= async (req, res)=>{
     const email = req.query.email;
 
     const orgsCollectionRef = database.collection('Organizers');
-    const orgDocSnapshot = await usersCollectionRef.where('email', '==', email).get();
+    const orgDocSnapshot = await orgsCollectionRef.where('email', '==', email).get();
     
     if (orgDocSnapshot.empty) {
-        res.status(500).json({
-            success: false,
+        res.status(404).json({
             error: 'No organization found with this email'
         })
     } else {
@@ -63,39 +58,57 @@ const readOrganizerByEmail=async (req, res)=>{
 const updateOrganizer=async (req, res)=>{
     const {id} = req.params;
     const orgRef = database.collection('Organizers').doc(id);
-    orgRef.update(req.body).then(() => {
-        res.json({
-            success: true
-        })
+
+    orgRef.get().then((orgDoc) => {
+        if (orgDoc.exists) {
+            orgRef.update(req.body).then(() => {
+                res.json({
+                    id: id
+                })
+            })
+        } else {
+            res.status(404).json({
+                error: 'Organizer could not be found'
+            })
+        }
     }).catch((error) => {
-        res.status(404).json({
-            success: false,
-            error: 'Organizer could not be found'
+        res.status(500).json({
+            error: error
         })
     })
 };
 
-const deleteOrganizer=async (req, res)=>{
+const deleteOrganizer= async (req, res)=>{
     const {id} = req.params;
     const orgRef = database.collection('Organizers').doc(id);
-    const orgDoc = orgRef.get();
-    orgRef.delete().then(() => {
-        res.json({
-            success: true
+    
+    orgRef.get().then((orgDoc) => {
+        if (orgDoc.exists) {
+            orgRef.delete().then(() => {
+                res.json({
+                    id: id
+                })
+            }).catch((error) => {
+                res.status(500).json({
+                    error: error
+                })
+            })
+        } else {
+            res.status(404).json({
+                error: 'Organizer does not exist'
+            })
+        }
+    }).catch((error) => {
+        res.status(500).json({
+            error: error
         })
-    }).catch((error) =>
-    res.status(404).json({
-        success: false,
-        error: 'Event could not be found'
     })
-    )
 };
 
-module.exports=
-{
-createOrganizer,
-readOrganizer,
-readOrganizerByEmail,
-updateOrganizer,
-deleteOrganizer,
+module.exports = {
+    createOrganizer,
+    readOrganizer,
+    readOrganizerByEmail,
+    updateOrganizer,
+    deleteOrganizer,
 };
